Sometimes we only need to remove pinboard tag
Pinboard bookmarks can have multiple tags. Since browser bookmark tags
are not exposed in the browser.bookmarks API, we model this by having
multiple browser bookmarks per one pinboard bookmark, each in the folder
named after tag. So, when we remove browser bookmark, it may be the case
that there are other bookmarks with the same url but in different
folders. If this is the case, we don't want to remove the pinboard
bookmark, but rather just modify it to have one less tag.